fix module3-task2 | add fixtures, refactor elements to functions, add and update templates, update constants
