+++ js/constants/questionTypes.js
@@ -1,0 +1,5 @@
+export default {
+  GUESS_EVERY_ITEM: 'Угадайте для каждого изображения фото или рисунок?',
+  GUESS_SINGLE_ITEM: 'Угадай, фото или рисунок?',
+  FIND_PAINT: 'Найдите рисунок среди изображений?'
+};
