+++ js/fixtures/rulesData.js
@@ -1,0 +1,7 @@
+const data = {
+  attempts: 10,
+  attemptTimeInSec: 30,
+  lives: 3
+};
+
+export default data;
