+++ js/templates/renderOptionChoose.js
